haggle/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
anthropic>=0.40.0
praw>=7.7.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
